requirements.txt
langchain
streamlit
langchain-google-genai
langchain-community
google-genai
python-docx
langchain-chroma
chromadb
pypdf
sentence-transformers

-e .
